Update on character chooser
A basic template on how the character will be selected has been applied.
Selection with "Arrow Keys" and "WASD" is available, but no visuals have
been applied.

Static class needs to be added to hold the details of each character, so
that the Choose Character script can be more structured.

Characters, can't use the same control scheme